backend/requirements-dev.txt
flake8==6.0.0
isort==5.12.0
mypy==1.0.1
mypy-extensions==1.0.0
pycodestyle==2.10.0
black==23.1.0
pyflakes==3.0.1
mccabe==0.7.0
click==8.1.3
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
tomli==2.0.1
typing_extensions==4.5.0
attrs==22.2.0
exceptiongroup==1.1.0
iniconfig==2.0.0
pluggy==1.0.0
pytest==7.2.1
django-stubs==1.15.0
django-stubs-ext==0.7.0
types-PyYAML==6.0.12.8
types-pytz==2022.7.1.2
Sphinx==5.3.0
sphinx-rtd-theme==1.2.0


python-dateutil~=2.8.1